+++ src/common/helpers.go
@@ -2,11 +2,13 @@ package common
 
 import (
 	"encoding/json"
+	"fmt"
 	"protocol"
 	"time"
 )
 
-func StringToSeriesArray(seriesString string) ([]*protocol.Series, error) {
+func StringToSeriesArray(seriesString string, args ...interface{}) ([]*protocol.Series, error) {
+	seriesString = fmt.Sprintf(seriesString, args...)
 	series := []*protocol.Series{}
 	err := json.Unmarshal([]byte(seriesString), &series)
 	return series, err