+++ CHANGELOG.md
@@ -116,8 +116,10 @@
 ## Features
 
 - [Issue #51](https://github.com/influxdb/influxdb/issues/51). Implement first and last aggregates
+- [Issue #35](https://github.com/influxdb/influxdb/issues/35). Support table aliases in Join Queries
 
 ## Bugfixes
 
+- Fix join and merges when the query is a descending order query
 - [Issue #57](https://github.com/influxdb/influxdb/issues/57). Don't panic when type of time != float
 - [Issue #63](https://github.com/influxdb/influxdb/issues/63). Aggregate queries should not have a sequence_number column